Visit nested properties directly when discovering properties
