+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/properties/annotations/AbstractOutputPropertyAnnotationHandler.java
@@ -42,7 +42,7 @@ public abstract class AbstractOutputPropertyAnnotationHandler extends AbstractPr
     }
 
     @Override
-    public void visitPropertyValue(String propertyName, PropertyValue value, PropertyMetadata propertyMetadata, PropertyVisitor visitor, BeanPropertyContext context) {
+    public void visitPropertyValue(String propertyName, PropertyValue value, PropertyMetadata propertyMetadata, PropertyVisitor visitor) {
         visitor.visitOutputFileProperty(propertyName, propertyMetadata.isAnnotationPresent(Optional.class), value, filePropertyType);
     }
 }
