+++ subprojects/model-core/src/main/java/org/gradle/internal/properties/annotations/NoOpPropertyAnnotationHandler.java
@@ -32,6 +32,6 @@ public class NoOpPropertyAnnotationHandler extends AbstractPropertyAnnotationHan
     }
 
     @Override
-    public void visitPropertyValue(String propertyName, PropertyValue value, PropertyMetadata propertyMetadata, PropertyVisitor visitor, BeanPropertyContext context) {
+    public void visitPropertyValue(String propertyName, PropertyValue value, PropertyMetadata propertyMetadata, PropertyVisitor visitor) {
     }
 }
